Split PageBuilder to PageBuilderImpl
